requirements.txt
flask
twilio
langchain
langchain-community
langchain-openai
openai
faiss-cpu
python-dotenv
tiktoken